ingest/requirements.txt
pdfplumber
sentence-transformers
supabase
python-dotenv